Added tests

+++ services/apatelet/store/store_test.go
@@ -53,7 +53,7 @@ func TestPollSingleTask(t *testing.T) {
 	// Poll single task and verify the timestamp is correct
 	retrieved, err := st.PeekTask()
 	assert.NoError(t, err)
-	assert.Equal(t, task.AbsoluteTimestamp, int32(retrieved))
+	assert.Equal(t, task.AbsoluteTimestamp, retrieved)
 
 	// Also verify it was not removed
 	assert.Equal(t, 1, st.LenTasks())
@@ -72,7 +72,7 @@ func TestMultipleTasks(t *testing.T) {
 	// Poll first task, which should be task 2
 	firstTaskTime, err := st.PeekTask()
 	assert.NoError(t, err)
-	assert.Equal(t, task2.AbsoluteTimestamp, int32(firstTaskTime))
+	assert.Equal(t, task2.AbsoluteTimestamp, firstTaskTime)
 
 	// Retrieve first two tasks
 	firstTask, err := st.PopTask()
@@ -87,7 +87,7 @@ func TestMultipleTasks(t *testing.T) {
 	lastTaskTime, err := st.PeekTask()
 	assert.NoError(t, err)
 	assert.Equal(t, 1, st.LenTasks())
-	assert.Equal(t, task3.AbsoluteTimestamp, int32(lastTaskTime))
+	assert.Equal(t, task3.AbsoluteTimestamp, lastTaskTime)
 }
 
 // TestUnsetFlag ensures the correct default value is returned for an unset flag (0), and an error is also returned
@@ -130,7 +130,7 @@ func TestArrayWithNil(t *testing.T) {
 	// Poll first task, which should be task 2
 	firstTaskTime, err := st.PeekTask()
 	assert.NoError(t, err)
-	assert.Equal(t, task2.AbsoluteTimestamp, int32(firstTaskTime))
+	assert.Equal(t, task2.AbsoluteTimestamp, firstTaskTime)
 
 	// Retrieve first task, and confirm it's task 2
 	firstTask, err := st.PopTask()
